beeper: fix .desktop launcher not working
The new Beeper's appimage has an installer part, where it places stuff in the user's home folder, including its own version of the .desktop file.
That .desktop file takes precedence over the one in the nix store, however it is broken as Exec is set to undefined, setting the APPIMAGE
env variable should solve that.